move batch workflowmanager + tests into their own files
